adressed canges to nnc_new_data_structure
