+++ libecl/CMakeLists.txt
@@ -93,6 +93,7 @@ foreach (name   ecl_alloc_cpgrid
                 ecl_util_make_date_no_shift
                 ecl_util_month_range
                 ecl_valid_basename
+                test_ecl_nnc_data
         )
         add_executable(${name} tests/${name}.c)
         target_link_libraries(${name} ecl)
@@ -164,9 +165,9 @@ target_link_libraries(ecl_nnc_export_get_tran ecl)
 add_test(NAME ecl_nnc_export_get_tran COMMAND ecl_nnc_export_get_tran
     ${_eclpath}/Troll/MSW_LGR/2BRANCHES-CCEWELLPATH-NEW-SCH-TUNED-AR3)
 
-add_executable(ecl_nnc_data tests/test_ecl_nnc_data.c)
-target_link_libraries(ecl_nnc_data ecl)
-add_test(NAME ecl_nnc_data COMMAND ecl_nnc_data
+add_executable(ecl_nnc_data_statoil_root tests/test_ecl_nnc_data_statoil_root.c)
+target_link_libraries(ecl_nnc_data_statoil_root ecl)
+add_test(NAME ecl_nnc_data_statoil_root COMMAND ecl_nnc_data_statoil_root
     ${_eclpath}/Troll/MSW_LGR/2BRANCHES-CCEWELLPATH-NEW-SCH-TUNED-AR3)
 
 add_executable(ecl_util_make_date_shift tests/ecl_util_make_date_shift.c)
